Adds Runner, Grid, and Car class
Grid contains an enum which inherits the same interface as the Car class so we can easily store both in the same 2D array to comprise the grid. Runner class for now just creates a grid and has a skeleton of the code it needs to create threads for the cars and the stoplights. I created a shitty quality grid just to test my enum and print methods for now.